requirements.txt
fastapi
uvicorn[standard]
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-chroma>=0.1.1
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
torch
pypdf>=4.2.0
python-dotenv
python-multipart>=0.0.20
pytesseract>=0.3.10
pdf2image>=1.17.0